devops-project-7188: Refactor arquitectura de asignaciones
