+++ src/services/historial.service.ts
@@ -1,6 +1,6 @@
-import { client } from "../../util/database_client";
-import { DatosInvalidos } from "../../util/errores";
-import { asignacionesModelo, DatosAsignacion } from "./asignaciones";
+import { client } from "../util/database_client";
+import { DatosInvalidos } from "../util/errores";
+import { DatosAsignacion } from "./asignaciones.service";
 
 export class HistorialAsignacionesModelo {
   async Crear(asignacion: DatosAsignacion) {
